+++ annotation-mongo/src/main/java/eu/europeana/annotation/mongo/service/PersistentAnnotationService.java
@@ -37,7 +37,7 @@ public interface PersistentAnnotationService extends AbstractNoSqlService<Persis
 	public List<? extends Annotation> getFilteredAnnotationList (
 			String europeanaId, String provider, String startOn, String limit, boolean isDisabled);
 
-	public PersistentAnnotation find(String europeanaId, String provider, Integer annotationNr);
+	public PersistentAnnotation find(String europeanaId, String provider, Long annotationNr);
 	
 	public PersistentAnnotation find(AnnotationId annoId);
 	
@@ -47,7 +47,7 @@ public interface PersistentAnnotationService extends AbstractNoSqlService<Persis
 	 * @param annotationNr
 	 * @throws AnnotationMongoRuntimeException - less or more than 1 object is found for the given arguments
 	 */
-	public void remove(String resourceId, String provider, Integer annotationNr);
+	public void remove(String resourceId, String provider, Long annotationNr);
 	
 	/**
 	 *
