+++ annotation-web/src/main/java/eu/europeana/annotation/web/service/controller/ManagementRest.java
@@ -53,7 +53,7 @@ public class ManagementRest extends BaseRest {
 				apiKey, "/admin/delete");
 			
 		try{
-			getAnnotationService().deleteAnnotation(resourceId, provider, Integer.valueOf(query));
+			getAnnotationService().deleteAnnotation(resourceId, provider, Long.valueOf(query));
 			response.success = true;
 		} catch (Exception e){
 			Logger.getLogger(SolrAnnotationConst.ROOT).error(e);
@@ -79,7 +79,7 @@ public class ManagementRest extends BaseRest {
 				apiKey, "/annotations/admin/index");
 			
 		try{
-			getAnnotationService().indexAnnotation(resourceId, provider, Integer.valueOf(query));
+			getAnnotationService().indexAnnotation(resourceId, provider, Long.valueOf(query));
 			response.success = true;
 		} catch (Exception e){
 			Logger.getLogger(SolrAnnotationConst.ROOT).error(e);
@@ -132,7 +132,7 @@ public class ManagementRest extends BaseRest {
 				apiKey, "/admin/disable");
 			
 		try{
-			getAnnotationService().disableAnnotation(resourceId, provider, Integer.valueOf(query));
+			getAnnotationService().disableAnnotation(resourceId, provider, Long.valueOf(query));
 			response.success = true;
 		} catch (Exception e){
 			Logger.getLogger(SolrAnnotationConst.ROOT).error(e);
